Override of delegate parameters in TextContentNode.swift to prevent crash
This is not cool, but, in the Library Texture ASTextNode.mm the method
gestureRecognizerShouldBegin is calling our delegate. It passes four
parameters none of which should be nil. However,
shouldLongPressLinkAttribute attribute: and value: are both nil. Since
the swift code defines these are non-optional it’s causing a crashing.
I changed the parameters to optional and not I get a warning because it
does not match the delegate prototype, however, we no longer get a
crash.